Refractored store plugin to CaseMixin and VariantMixin
