+++ puzzle/plugins/sql/store.py
@@ -3,7 +3,6 @@
 puzzle.plugins.sql.store
 ~~~~~~~~~~~~~~~~~~
 """
-import itertools
 import logging
 import os
 
@@ -16,6 +15,9 @@ from puzzle.models import Individual as BaseIndividual
 from puzzle.models.sql import (BASE, Case, Individual, PhenotypeTerm, GeneList,
                                CaseGenelistLink, Resource)
 from puzzle.plugins import VcfPlugin, Plugin
+
+from . import VariantMixin, CaseMixin
+
 try:
     from puzzle.plugins import GeminiPlugin
 except ImportError as e:
@@ -25,7 +27,7 @@ from puzzle.utils import hpo_genes
 logger = logging.getLogger(__name__)
 
 
-class Store(Plugin):
+class Store(VariantMixin, CaseMixin, Plugin):
 
     """SQLAlchemy-based database object.
     .. note::
@@ -124,113 +126,6 @@ class Store(Plugin):
         self.session.flush()
         self.session.commit()
         return self
-
-    def add_case(self, case_obj, vtype='snv', mode='vcf', ped_svg=None):
-        """Load a case with individuals.
-
-        Args:
-            case_obj (puzzle.models.Case): initialized case model
-        """
-        new_case = Case(case_id=case_obj.case_id,
-                        name=case_obj.name,
-                        variant_source=case_obj.variant_source,
-                        variant_type=vtype,
-                        variant_mode=mode,
-                        pedigree=ped_svg)
-
-        # build individuals
-        inds = [Individual(
-            ind_id=ind.ind_id,
-            name=ind.name,
-            mother=ind.mother,
-            father=ind.father,
-            sex=ind.sex,
-            phenotype=ind.phenotype,
-            ind_index=ind.ind_index,
-            variant_source=ind.variant_source,
-            bam_path=ind.bam_path,
-        ) for ind in case_obj.individuals]
-
-        new_case.individuals = inds
-        self.session.add(new_case)
-        self.save()
-        return new_case
-
-    def delete_case(self, case_obj):
-        """Delete a case from the database
-
-        Args:
-            case_obj (puzzle.models.Case): initialized case model
-        """
-        for ind_obj in case_obj.individuals:
-            self.delete_individual(ind_obj)
-        logger.info("Deleting case {0} from database".format(case_obj.case_id))
-        self.session.delete(case_obj)
-        self.save()
-        return case_obj
-
-    def delete_individual(self, ind_obj):
-        """Delete a case from the database
-
-        Args:
-            ind_obj (puzzle.models.Individual): initialized individual model
-        """
-        logger.info("Deleting individual {0} from database".format(ind_obj.ind_id))
-        self.session.delete(ind_obj)
-        self.save()
-        return ind_obj
-
-    def case(self, case_id):
-        """Fetch a case from the database."""
-        case_obj = self.query(Case).filter_by(case_id=case_id).first()
-        if case_obj is None:
-            case_obj = BaseCase(case_id='unknown')
-        return case_obj
-
-    def individual(self, ind_id):
-        """Fetch a case from the database."""
-        ind_obj = self.query(Individual).filter_by(ind_id=ind_id).first()
-        if ind_obj is None:
-            ind_obj = BaseIndividual(ind_id='unknown')
-        return ind_obj
-
-    def cases(self):
-        """Fetch all cases from the database."""
-        return self.query(Case)
-
-    def get_individuals(self, ind_ids=None):
-        """Fetch all individuals from the database."""
-        query = self.query(Individual)
-        if ind_ids:
-            query = query.filter(Individual.ind_id.in_(ind_ids))
-        return query
-
-    def variants(self, case_id, skip=0, count=30, filters=None):
-        """Fetch variants for a case."""
-        filters = filters or {}
-        logger.debug("Fetching case with case_id:{0}".format(case_id))
-        case_obj = self.case(case_id)
-        plugin, case_id = self.select_plugin(case_obj)
-        self.filters = plugin.filters
-
-        gene_lists = (self.gene_list(list_id) for list_id
-                      in filters.get('gene_lists', []))
-        nested_geneids = (gene_list.gene_ids for gene_list in gene_lists)
-        gene_ids = set(itertools.chain.from_iterable(nested_geneids))
-
-        if filters.get('gene_ids'):
-            filters['gene_ids'].extend(gene_ids)
-        else:
-            filters['gene_ids'] = gene_ids
-        variants = plugin.variants(case_id, skip, count, filters)
-        return variants
-
-    def variant(self, case_id, variant_id):
-        """Fetch a single variant from variant source."""
-        case_obj = self.case(case_id)
-        plugin, case_id = self.select_plugin(case_obj)
-        variant = plugin.variant(case_id, variant_id)
-        return variant
 
     def add_phenotype(self, ind_obj, phenotype_id):
         """Add a phenotype term to the case."""
